Changing how the waveform is colored
The Siri-style waveform now consumes SCSS variables in
`PreviewContainerComponent`, which then gets passed into the `Animator`
object and through to the realtime waveform renderer.

I added a shading function so it automatically calculates lighter
versions of the base wave color. The alternative was passing in 3 sass
variables, two of them calculated via `lighten`, but I decided this was
a little cleaner and put less cruft in the `_variables.scss` file.