Twitter posting via server instead  of Lambda
This moves all the Twitter posting to the regular Shortcut server via a
`social-post` POST endpoint with CORS enabled. POSTing to that endpoint
with data formatted as, example:

```json
{"video_data":{"Bucket":"shortcut-perf-test","Key":"580_1.58_2.16_1f97c9.mp4","message":"Successfully uploaded to S3.","url":"https://shortcut-perf-test.s3-us-west-2.amazonaws.com/580_1.58_2.16_1f97c9.mp4"},"msg":"Testing this out.","twitter_info":"{\"accessToken\":\"1475905308-WDNvkYvUXuUIHT5XlYQbstS4Erb54asMTPflB80\",\"accessTokenSecret\":\"E1n4dSduwvVkzP1IlShHezJVrX2EmMYkkjD0z8vBE4FQD\",\"userId\":\"1475905308\"}","facebook_info":false}
```

will provide the server with the twitter credentials to post to Twitter
on behalf of the user. The server then doenloads the mp4 from s3, and
uploads it to Twitter with the status message `msg` specified by the
user.