Remove mySQL DB as dependency
This commit removes all the mySQL DB stuff which includes tracking
metrics on shared items and also episode versioning. We may put this
back in a future build but it's just complicating things to keep it
around.

As a result of removing this functionality, I had to rewrite the way
episode data is loaded on the landing page. In the process of doing
this, I noticed that the dev React server was not pulling data from the
actual episodes list, so this fixes that. Now, when you're running the
client from the dev front-end server, it hotloads `episodes.json` from
the dataBucket.

Also removing verification headers, which were added to mitigate web
scraping. Web scraping is a good part of a healthy open web. If
individual implementors want to maintain a branch that supports this, by
all means do so, but also consider that scrapers for the most part help
with things like visibility for your podcast.

Closes #48.