Changing path to static assets
Since both client and server are in the same repo, we now know exactly
where the static assets are being built to in the client, relative to
the server. This changes the express `static` server to point directly
to `../client/dist/` which removes a step in the compilation/test
process.