Updating menu UI
There were some bugs with rendering menu UI in several different views.
This updates some text color for the mobile menu and removes an unneeded
component from the desktop menu as well as fixing the general layout of
the desktop menu. (We had MenuItem wrapping LoginComponent, which it
turns out was totally unneeded. Hooray for one fewer component!)

Closes #70.