Removing S3 bucket region from clientside config
There was some confusion from users around finding the correct region
corresponding to the S3 bucket. Amazon does not make this easy.
Fortunately, there is a way for the client to access the contents of a
bucket with just the bucket name, rather than a region. This commit
changes our S3 path to the regionless kind, so the `s3Region` variable
no longer needs to be set in `base.js` on the client.