Serve the podcast RSS feed from Shortcut server
This enables us to get around any CORS issues with the podcast RSS feed.
We just re-serve it from our server with CORS already enabled.

By parsing the RSS serverside and serving it up as plain json, we
no longer have to include a costly XML parsing step on the client. #80