Server and client can use RSS feed as data source
There is now the option to point shortcut's client and server to an RSS
feed URL instead of maintaining an `episodes.json` file with manual
metadata. The admin sets the `useRSS` option in
client/src/config/default.js to `true` and provides a feed URL in the
server's .env file `RSS_FEED`.

At every point in the code where `episodes.json` might be accessed for
metadata about all episodes, the code now branches if RSS is enabled and
instead parses the RSS feed.

This closes #82.