Admin panel
There is now an admin panel that can be reached at the `/#/admin` route.
The panel shows a list of episodes that Shortcut knows about, as well as
their state as enabled or disabled (disabled by default). This state is
stored in `server/adminData.json` on the server.

This closes #81, because we no longer need a specific user-defined GUID
to say "episode 3098281ABD23 is enabled" -- now we use the `guid` tag
defined in the rss spec.