Add nodemon to shortcut-server devDependencies
Documentation and package.json provide guidance to using nodemon to ease
the development process. However, it was missing from devDependencies in
server/package.json and a manual install instruction in server/README.md

As npm has functionality to reflect dependencies that should not be run in
--production (or when the NODE_ENV environment variable is set to production)
then let's utilize that.

This reduces the number of manual commands that must be run to setup the
recommended shortcut development environment.