Remove infinite scroll on landing page
Infinite scroll is overengineering for this particular UX element. Even
the weakest smartphones can render ~1000 plaintext divs no problem, and
while getting all the way down to the bottom might be a problem, as long
as we don't provide a search feature (which is forthcoming) at least
rendering all the episodes on one landing page lets a user ctrl+F or
"find in page" on mobile to find the particular episode they are looking
for. Even pagination (the usual alternative to infinite scroll) would
make this impossible.