Removing the '/rss' endpoint
Per @therewasaguy's
[review](https://github.com/FeelTrainCoop/shortcut/pull/96#discussion_r168960372)
the '/rss' and '/recent' endpoints return the same thing if you're set
to RSS mode. So no reason to have a dedicated '/rss' endpoint, and also
no more need for a `useRSS` configuration variable on the client.

The client just asks the server for recent episodes, and the server
knows if it needs to retrieve those from the RSS feed or from the JSON
file.