Admin will always update episode cache
Accessing the admin panel will update the cache locally. It's just a
GET request to the RSS feed (in the case of RSS), which we wouldn't want
a normal user to intialize but is fine if an admin does it. The server
endpoint is behind an auth gate so only an admin will ever kick it off.