Tests passing
In preparation for setting up basic travis-ci integration, I've fixed up
some of the basic tests. ShareContainer and NavBar return a styles
object, not the component itself, so for now we're just going to test
that they return the styles object as material-ui wants them to. Later,
we will have to of course expose the actual component to do real
testing.