Add initial Libsyn handshake from setup wizard
This closes #107. On going to the setup wizard page, the user is
presented with a prompt asking for a Libsyn URL. When that's given, we
send it to the server. The server sets the "episodeSource" in the DB to
be of type "libsyn" with the RSS url of the libsyn podcast. We change
the routing of the RSS parsing so that it defaults to the RSS feed
pointed to in "episodeSource" if that key exists, otherwise it uses the
RSS_FEED variable set in .env like normal.

We also add a new helper function on the server that parses the metadata
for show from an rss feed (previously we only had a function to get
individual episode data).