Admin user/pass configuration in setup wizard
We now POST and store a username/password as part of the Setup Wizard
process. This commit adds the POST endpoint to the server, adds the
clientside support in the wizard for sending the data. There is some
basic clientside form validation. In theory we could do serverside
validation but these endpoints are locked down, the only users with
access to them (uncredentialed) are the users setting up credentials in
the first place.

We use `bcrypt-nodejs` to store hashed username/password in sqlite. If
an old-style password is specified in the .env file and the user hasn't
set the password through the wizard, we store that password as well
(with the default username 'admin').

Closes #108.