Convert from flat-cache to sqlite (#113)
* Move our cache implementation to sqlite/sqlcipher

This commit removes the `flat-cache` package from Shortcut in favor of
`sqlite3` built with `sqlcipher`. We were using the `flat-cache` package
to persist simple key/value data in a JSON file for keeping track of
which episodes have been enabled/disabled from visibility in the main
episode listing via the admin panel. Now we use a key/value store
stored in a `kvs` table in sqlite, but also the episode state is stored
in an "episodes" table that is not a key/value store. We may extend this
table in the future, we may not, but what it means is that sqlite is
searching the table rather than us pulling the entire table in Node.js
and running our own search.

Functionality remains identical to how it was before this commit.