Sanitize and send correct inputs to setEpisode
We need to be careful when setting stuff in sqlite -- there's no boolean
type, so we store booleans as an integer 1 or 0, as they recommend in
their official documentation. But we were sending the string "true" and
"false" to the server, which messed everything up.

This commit makes the client send "0" or "1" and we convert literally
everything that is not the string "1" to 0, and "1" is 1. So the only
correct way to enabled an episode on this endpoint is to send

`?enabled=1`

as a url parameter.

Closes #114.