Create a Libsyn Setup Wizard (#116)
* Move our cache implementation to sqlite/sqlcipher

This commit removes the `flat-cache` package from Shortcut in favor of
`sqlite3` built with `sqlcipher`. We were using the `flat-cache` package
to persist simple key/value data in a JSON file for keeping track of
which episodes have been enabled/disabled from visibility in the main
episode listing via the admin panel. Now we use a key/value store
stored in a `kvs` table in sqlite, but also the episode state is stored
in an "episodes" table that is not a key/value store. We may extend this
table in the future, we may not, but what it means is that sqlite is
searching the table rather than us pulling the entire table in Node.js
and running our own search.

Functionality remains identical to how it was before this commit.

* Add sqlite DB to gitignore

* Add a basic landing page for the setup wizard

* Add initial Libsyn handshake from setup wizard

This closes #107. On going to the setup wizard page, the user is
presented with a prompt asking for a Libsyn URL. When that's given, we
send it to the server. The server sets the "episodeSource" in the DB to
be of type "libsyn" with the RSS url of the libsyn podcast. We change
the routing of the RSS parsing so that it defaults to the RSS feed
pointed to in "episodeSource" if that key exists, otherwise it uses the
RSS_FEED variable set in .env like normal.

We also add a new helper function on the server that parses the metadata
for show from an rss feed (previously we only had a function to get
individual episode data).

* Admin user/pass configuration in setup wizard

We now POST and store a username/password as part of the Setup Wizard
process. This commit adds the POST endpoint to the server, adds the
clientside support in the wizard for sending the data. There is some
basic clientside form validation. In theory we could do serverside
validation but these endpoints are locked down, the only users with
access to them (uncredentialed) are the users setting up credentials in
the first place.

We use `bcrypt-nodejs` to store hashed username/password in sqlite. If
an old-style password is specified in the .env file and the user hasn't
set the password through the wizard, we store that password as well
(with the default username 'admin').

Closes #108.

* Sanitize and send correct inputs to setEpisode

We need to be careful when setting stuff in sqlite -- there's no boolean
type, so we store booleans as an integer 1 or 0, as they recommend in
their official documentation. But we were sending the string "true" and
"false" to the server, which messed everything up.

This commit makes the client send "0" or "1" and we convert literally
everything that is not the string "1" to 0, and "1" is 1. So the only
correct way to enabled an episode on this endpoint is to send

`?enabled=1`

as a url parameter.

Closes #114.

* Improvements to setup flow and security

* creates a new public-viewable API that lets you know if a data source
has been set on shortcut
* the setup page now checks for that data source set and warns you if
you try to set up a shortcut instance with already set data
* setup now require credentials except on first setup (which you can
just press "ok" to get through the password prompts)
* wipe the `episodes` table clean on a new setup

Addresses most of #112.

* Adding missing API route

* Add password confirmation field to setup form

Addresses #112.