Further updates for React Router v4
There were a bunch of issues with hashHistory, since that no longer
exists.

This commit also removes our entire "view" system that was really hacked
together and also didn't perform very well. We now rely on strict
react-router-dom routes, which is a huge improvement in code readability
(it's declarative!) and performance.

Closes #12. Closes #13.