Encode URI as episode number
Internally we use the guid, which is often a URI, as the episode number
if there is no episode number provided explicitly in the RSS feed. If
this isn't encoded, then when it's used as a routing parameter in a
web application URL, it can mess up the routing. Encoding it fixes
this problem.